module_search: Handle overflow
Previous this was guarded aginst by conditions in gap_common.
It's better to just handle as module_search_euclid also uses int128
so only fraction slower and never have to worry.

+++ modulo_search.cpp
@@ -180,9 +180,7 @@ uint64_t modulo_search_euclid_gcd2(
         uint64_t M, uint64_t D, uint64_t max_m, uint64_t SL,
         uint64_t prime, uint64_t base_r) {
 
-    // NOTE: this overflows with large M, prime
-    // This should be guarded by gap_common.argparse
-    uint64_t modulo = (base_r * M + SL) % prime;
+    uint64_t modulo = ((__int128) base_r * M + SL) % prime;
     uint64_t init = modulo;
     uint32_t two_SL = SL << 1;
 
@@ -232,9 +230,7 @@ uint64_t modulo_search_euclid_gcd(
         uint64_t prime, uint64_t base_r) {
     uint64_t mi = 0;
 
-    // NOTE: this overflows with large M, prime
-    // This should be guarded by gap_common.argparse
-    uint64_t modulo = (base_r * M) % prime;
+    uint64_t modulo = ((__int128) base_r * M) % prime;
     while (mi < max_m) {
         if ( (modulo <= SL) || (modulo + SL) >= prime) {
             if (gcd(M + mi, D) > 1) {
